# openapi.yml
openapi: 3.0.3
info:
  title: Processing Service API
  description: |
    **Processing Service** (Consumer → Ollama Orchestrator + Aggregator).

    Сервис читает страницы из Kafka (`raw_chunks`), сохраняет распознанный текст в
    PostgreSQL, выполняет детерминированный иерархический diff полных документов,
    передаёт найденные области в локальную LLM только для строгой проверки и
    доставляет финальный JSON на фронтенд через WebSocket.

    **Пайплайн (в связке с Chunking & Producer):**
    1. Фронтенд → `POST /api/compare` на Chunking Service (`:5000`) — загрузка файлов, `job_id`
    2. Chunking → Kafka `raw_chunks`
    3. **Processing Service** (`:5001`) — OCR → PostgreSQL → hierarchical diff → LLM verify
    4. Фронтенд → WebSocket `ws://localhost:5001/ws/jobs/{job_id}` — прогресс и результат
    5. (fallback) `GET /api/jobs/{job_id}/result` — получить результат без WebSocket

    **Swagger UI:** `http://localhost:5001/docs`

    **WebSocket** не отображается в стандартном Swagger UI; контракт описан в разделе
    `x-websocket` и в схеме `WebSocketEvent`.
  version: 2.0.0
  contact:
    name: Processing Service

servers:
  - url: http://localhost:5001
    description: Локальный Processing Service

tags:
  - name: Health
    description: Проверка доступности сервиса, Kafka, PostgreSQL и Ollama
  - name: Jobs
    description: Статус задач и финальный результат сравнения

paths:
  /health:
    get:
      tags: [Health]
      summary: Проверка состояния сервиса
      description: |
        Возвращает общий статус и доступность Kafka (consumer + producer + aggregator relay),
        PostgreSQL и Ollama API.
      operationId: health
      responses:
        "200":
          description: Сервис отвечает
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                ok:
                  summary: Всё доступно
                  value:
                    status: ok
                    kafka_connected: true
                    ollama_reachable: true
                    postgres_connected: true
                degraded:
                  summary: Частичная недоступность
                  value:
                    status: degraded
                    kafka_connected: false
                    ollama_reachable: true
                    postgres_connected: true

  /api/jobs:
    get:
      tags: [Jobs]
      summary: Список активных задач
      description: |
        Возвращает задачи из PostgreSQL. Список сохраняется при перезапуске сервиса.
      operationId: listJobs
      responses:
        "200":
          description: Список задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobStatusResponse"
              example:
                - job_id: "550e8400-e29b-41d4-a716-446655440000"
                  document_id: "550e8400-e29b-41d4-a716-446655440000"
                  status: processing
                  processed_chunks: 3
                  total_chunks: 10
                  message: "Анализ chunk 3/10..."
    post:
      tags: [Jobs]
      summary: Ранняя регистрация задачи
      operationId: registerJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobRegisterRequest"
      responses:
        "200":
          description: Задача зарегистрирована в PostgreSQL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"

  /api/jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Статус задачи по job_id
      description: |
        Возвращает сохранённый в PostgreSQL прогресс OCR и этап сравнения.
        Для live-обновлений предпочтительнее WebSocket.
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Статус задачи
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
              example:
                job_id: "550e8400-e29b-41d4-a716-446655440000"
                document_id: "550e8400-e29b-41d4-a716-446655440000"
                status: processing
                processed_chunks: 7
                total_chunks: 10
                message: "Анализ chunk 7/10..."
        "404":
          description: Задача не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Job 550e8400-e29b-41d4-a716-446655440000 not found"

  /api/jobs/{job_id}/ocr:
    get:
      tags: [Jobs]
      summary: Просмотр сохранённого OCR
      description: |
        Возвращает точный текст всех страниц/чанков, сохранённый в PostgreSQL
        до запуска иерархического diff. Параметр `side` ограничивает ответ
        первым или вторым документом.
      operationId: getJobOcr
      parameters:
        - $ref: "#/components/parameters/JobId"
        - name: side
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2
      responses:
        "200":
          description: Сохранённые OCR-чанки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OcrJobResponse"
        "404":
          description: Задача не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/jobs/{job_id}/result:
    get:
      tags: [Jobs]
      summary: Финальный результат сравнения (REST fallback)
      description: |
        Возвращает результат полного иерархического сравнения.
        Доступен после публикации единственного full-document `processed_results`.

        Рекомендуемый способ получения результата — WebSocket (`type: result`).
        Этот endpoint — fallback при обрыве WebSocket или для polling.
      operationId: getJobResult
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Результат сравнения готов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultResponse"
              examples:
                withDifferences:
                  summary: Есть расхождения
                  value:
                    comparison:
                      identical: false
                      differences:
                        - line_number: 5
                          file1_line: "Сумма: 100 000 руб."
                          file2_line: "Сумма: 120 000 руб."
                        - line_number: 12
                          file1_line: "Срок: 30 дней"
                          file2_line: "Срок: 45 дней"
                identical:
                  summary: Документы идентичны
                  value:
                    comparison:
                      identical: true
                      differences: []
        "404":
          description: Результат ещё не готов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Result for job 550e8400-e29b-41d4-a716-446655440000 is not ready yet"

x-websocket:
  /ws/jobs/{job_id}:
    summary: Live-прогресс и финальный результат
    description: |
      Подключение фронтенда к задаче сравнения.

      **URL:** `ws://localhost:5001/ws/jobs/{job_id}`

      **Когда подключаться:** сразу после получения `job_id` от Chunking Service
      (`POST /api/compare` на порту 5000).

      **Поведение сервера:**
      - При подключении, если результат уже готов, сразу отправляется событие `result`
      - В процессе обработки приходят события `status` из Kafka topic `status_updates`
      - По завершении агрегации приходит событие `result` с финальным JSON
      - Клиент может отправлять любой текст (ping) для поддержания соединения

      **Типы событий (входящие сообщения):** `status`, `result`, `error`
    parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ID задачи из ответа Chunking Service
    messages:
      status:
        summary: Обновление прогресса
        payload:
          $ref: "#/components/schemas/WebSocketStatusEvent"
      result:
        summary: Финальный результат сравнения
        payload:
          $ref: "#/components/schemas/WebSocketResultEvent"
      error:
        summary: Ошибка обработки
        payload:
          $ref: "#/components/schemas/WebSocketErrorEvent"

x-kafka:
  description: |
    Внутренние Kafka-топики Processing Service (для интеграции с Chunking & Producer).
    HTTP API для этих топиков нет — только документация контракта.
  consumes:
    - topic: raw_chunks
      description: Входящие чанки от Chunking & Producer
      schema:
        $ref: "#/components/schemas/RawChunkMessage"
      key: job_id
  produces:
    - topic: processed_results
      description: Один финальный результат иерархического diff на задачу
      schema:
        $ref: "#/components/schemas/ProcessedResultMessage"
      key: job_id
    - topic: status_updates
      description: Статусы для фронтенда (ретранслируются в WebSocket)
      schema:
        $ref: "#/components/schemas/StatusUpdateMessage"
      key: job_id
    - topic: raw_chunks_dlt
      description: Dead Letter Topic — сообщения с ошибками валидации/обработки
      key: job_id

components:
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      description: UUID задачи сравнения (из Chunking Service)
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

  schemas:
    HealthResponse:
      type: object
      required: [status, kafka_connected, ollama_reachable, postgres_connected]
      properties:
        status:
          type: string
          enum: [ok, degraded]
          description: |
            `ok` — Kafka, PostgreSQL и Ollama доступны;
            `degraded` — хотя бы один компонент недоступен
        kafka_connected:
          type: boolean
          description: Producer, raw_chunks consumer и aggregator relay подключены
        ollama_reachable:
          type: boolean
          description: Ollama API отвечает на GET /api/tags
        postgres_connected:
          type: boolean
          description: PostgreSQL отвечает и OCR-хранилище доступно

    JobStatusResponse:
      type: object
      required:
        - job_id
        - document_id
        - status
        - processed_chunks
        - total_chunks
        - message
      properties:
        job_id:
          type: string
          format: uuid
        document_id:
          type: string
          description: ID документа (обычно совпадает с job_id)
        status:
          type: string
          enum: [queued, preparing, processing, ocr_ready, comparing, completed, failed]
          description: |
            `processing` — идёт OCR;
            `ocr_ready` — весь OCR сохранён;
            `comparing` — выполняется иерархический diff и LLM-проверка;
            `completed` — результат готов;
            `failed` — ошибка обработки
        processed_chunks:
          type: integer
          minimum: 0
          description: Число обработанных чанков
        total_chunks:
          type: integer
          minimum: 0
          description: Общее число чанков в задаче
        message:
          type: string
          description: Человекочитаемое сообщение о текущем шаге
          example: "Анализ chunk 3/10..."

    JobRegisterRequest:
      type: object
      required: [job_id, user_id]
      properties:
        job_id:
          type: string
        user_id:
          type: string
          format: uuid
        document_id:
          type: string
          nullable: true
        total_chunks:
          type: integer
          minimum: 0
          default: 0
        status:
          type: string
          default: queued
        message:
          type: string
          default: "Ожидание чанков из Kafka..."
        file1_name:
          type: string
          default: ""
        file2_name:
          type: string
          default: ""

    OcrChunkResponse:
      type: object
      required:
        - chunk_index
        - side
        - filename
        - format
        - source_content_type
        - was_ocr
        - is_missing
        - text_content
        - created_at
        - updated_at
      properties:
        chunk_index:
          type: integer
          minimum: 1
        side:
          type: integer
          minimum: 1
          maximum: 2
        filename:
          type: string
        format:
          type: string
        source_content_type:
          type: string
          enum: [text, image, missing]
        was_ocr:
          type: boolean
        is_missing:
          type: boolean
        text_content:
          type: string
          description: Полный текст страницы/чанка до сравнения
        ocr_model:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OcrJobResponse:
      type: object
      required:
        - job_id
        - document_id
        - status
        - total_chunks
        - processed_chunks
        - chunks
      properties:
        job_id:
          type: string
        document_id:
          type: string
        status:
          type: string
        total_chunks:
          type: integer
        processed_chunks:
          type: integer
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/OcrChunkResponse"

    LineDifference:
      type: object
      properties:
        candidate_id:
          type: string
          nullable: true
          description: Стабильный ID кандидата внутри запуска сравнения
        line_number:
          type: integer
          nullable: true
          description: Номер строки (нумерация с 1)
        file1_line:
          type: string
          nullable: true
          maxLength: 200
          description: Фрагмент строки из первого документа (макс. 200 символов)
        file2_line:
          type: string
          nullable: true
          maxLength: 200
          description: Фрагмент строки из второго документа (макс. 200 символов)
        file1_span:
          type: array
          nullable: true
          minItems: 2
          maxItems: 2
          items:
            type: integer
          description: |
            [start, end] — индексы символов внутри file1_line (0-based, start включительно, end исключительно).
            Для подсветки на фронте: file1_line.slice(start, end)
          example: [12, 28]
        file2_span:
          type: array
          nullable: true
          minItems: 2
          maxItems: 2
          items:
            type: integer
          description: |
            [start, end] — индексы символов внутри file2_line (0-based, start включительно, end исключительно).
            Для подсветки на фронте: file2_line.slice(start, end)
          example: [12, 30]
        category:
          type: string
          enum: [substantive, technical, alignment_error, ocr_uncertain]
          default: substantive
        technical_type:
          type: string
          nullable: true
          enum:
            - markdown
            - numbering
            - page_number
            - list_marker
            - header_footer
            - dash
            - line_wrap
        reason:
          type: string
          nullable: true
        confidence:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 1
        protection_tags:
          type: array
          items:
            type: string
        file1_page:
          type: integer
          nullable: true
          minimum: 1
        file2_page:
          type: integer
          nullable: true
          minimum: 1
        file1_block:
          type: integer
          nullable: true
          minimum: 1
        file2_block:
          type: integer
          nullable: true
          minimum: 1
        file1_source_type:
          type: string
          nullable: true
        file2_source_type:
          type: string
          nullable: true

    ComparisonResult:
      type: object
      required: [identical, verdict, differences]
      properties:
        identical:
          type: boolean
          description: true — отсутствуют различия любого типа
        verdict:
          type: string
          enum: [identical, content_equal, different]
          description: Итог с отдельным состоянием совпадения содержимого
        differences:
          type: array
          items:
            $ref: "#/components/schemas/LineDifference"

    ResultResponse:
      type: object
      required: [comparison]
      properties:
        comparison:
          $ref: "#/components/schemas/ComparisonResult"

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Описание ошибки (формат FastAPI HTTPException)

    WebSocketEvent:
      type: object
      required: [type, job_id, data]
      properties:
        type:
          type: string
          enum: [status, result, error]
        job_id:
          type: string
          format: uuid
        data:
          type: object
          description: Полезная нагрузка (зависит от type)

    WebSocketStatusEvent:
      allOf:
        - $ref: "#/components/schemas/WebSocketEvent"
        - type: object
          properties:
            type:
              type: string
              enum: [status]
            data:
              $ref: "#/components/schemas/StatusUpdateMessage"

    WebSocketResultEvent:
      allOf:
        - $ref: "#/components/schemas/WebSocketEvent"
        - type: object
          properties:
            type:
              type: string
              enum: [result]
            data:
              type: object
              required: [comparison]
              properties:
                comparison:
                  $ref: "#/components/schemas/ComparisonResult"

    WebSocketErrorEvent:
      allOf:
        - $ref: "#/components/schemas/WebSocketEvent"
        - type: object
          properties:
            type:
              type: string
              enum: [error]
            data:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                details:
                  type: object
                  additionalProperties: true

    StatusUpdateMessage:
      type: object
      description: Payload в Kafka topic status_updates и в WebSocket type=status
      required:
        - job_id
        - document_id
        - status
        - processed_chunks
        - total_chunks
        - message
      properties:
        job_id:
          type: string
          format: uuid
        document_id:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed]
        processed_chunks:
          type: integer
          minimum: 0
        total_chunks:
          type: integer
          minimum: 1
        message:
          type: string
          example: "Анализ chunk 5/10..."
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 UTC

    RawChunkMessage:
      type: object
      description: Входящее сообщение из Kafka topic raw_chunks (от Chunking Service)
      required: [job_id, document_id, chunk_index, total_chunks]
      properties:
        job_id:
          type: string
          format: uuid
          description: Ключ сообщения Kafka
        document_id:
          type: string
        chunk_index:
          type: integer
          minimum: 1
          description: Номер чанка (1..total_chunks)
        total_chunks:
          type: integer
          minimum: 1
          description: Одинаковое значение во всех сообщениях задачи
        file1:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ChunkContent"
        file2:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ChunkContent"

    ChunkContent:
      type: object
      required: [filename, format, content_type, content]
      properties:
        filename:
          type: string
        format:
          type: string
          enum: [pdf, docx]
        content_type:
          type: string
          enum: [text, image]
          description: text — для DOCX; image — base64 PNG для PDF-страниц
        content:
          type: string
          description: Текст чанка или base64-encoded PNG

    ProcessedResultMessage:
      type: object
      description: |
        Единственное full-document сообщение в Kafka topic processed_results
        (`chunk_index=1`, `total_chunks=1`).
      required: [job_id, document_id, chunk_index, total_chunks, ollama]
      properties:
        job_id:
          type: string
          format: uuid
        document_id:
          type: string
        chunk_index:
          type: integer
        total_chunks:
          type: integer
        ollama:
          type: object
          description: Сводка LLM-проверки детерминированных diff-областей
        comparison_fragment:
          type: object
          nullable: true
          description: Итог алгоритма с подтверждёнными буквальными различиями
          allOf:
            - $ref: "#/components/schemas/ComparisonResult"
        processed_at:
          type: string
          format: date-time